Move tests related to handles to their own module
